# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Note {
  id: number;
  title: string;
  content: string;
  tags: string;
  /** @nullable */
  tldr?: string | null;
  /** @nullable */
  keyConcepts?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNoteInput {
  title: string;
  content: string;
  tags?: string;
  /** @nullable */
  tldr?: string | null;
  /** @nullable */
  keyConcepts?: string | null;
}

export interface UpdateNoteInput {
  title?: string;
  content?: string;
  tags?: string;
  /** @nullable */
  tldr?: string | null;
  /** @nullable */
  keyConcepts?: string | null;
}

export interface Flashcard {
  id: number;
  /** @nullable */
  noteId?: number | null;
  question: string;
  answer: string;
  /** @nullable */
  lastReviewed?: string | null;
  /** @nullable */
  nextReview?: string | null;
  difficulty: number;
  createdAt: string;
}

export interface CreateFlashcardInput {
  /** @nullable */
  noteId?: number | null;
  question: string;
  answer: string;
}

export interface UpdateFlashcardInput {
  lastReviewed?: string;
  nextReview?: string;
  difficulty?: number;
}

export interface FlashcardItem {
  question: string;
  answer: string;
}

export interface Quiz {
  id: number;
  /** @nullable */
  noteId?: number | null;
  title: string;
  /** @nullable */
  score?: number | null;
  total: number;
  createdAt: string;
}

export interface CreateQuizInput {
  /** @nullable */
  noteId?: number | null;
  title: string;
  score?: number;
  total: number;
}

export type QuizQuestionType =
  (typeof QuizQuestionType)[keyof typeof QuizQuestionType];

export const QuizQuestionType = {
  multiple_choice: "multiple_choice",
  true_false: "true_false",
} as const;

export interface QuizQuestion {
  question: string;
  type: QuizQuestionType;
  options: string[];
  answer: string;
  explanation: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  content: string;
  /** @nullable */
  subject?: string | null;
  createdAt: string;
}

export type CreateChatMessageInputRole =
  (typeof CreateChatMessageInputRole)[keyof typeof CreateChatMessageInputRole];

export const CreateChatMessageInputRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface CreateChatMessageInput {
  role: CreateChatMessageInputRole;
  content: string;
  /** @nullable */
  subject?: string | null;
}

export interface StudyPlan {
  id: number;
  subjects: string;
  examDate: string;
  planData: string;
  createdAt: string;
}

export interface CreateStudyPlanInput {
  subjects: string;
  examDate: string;
  planData: string;
}

export interface ProgressStats {
  notesCount: number;
  flashcardsStudied: number;
  quizzesCompleted: number;
  studyStreak: number;
  totalFlashcards: number;
}

export interface ChatMessageItem {
  role: string;
  content: string;
}

export interface AiChatInput {
  message: string;
  subject?: string;
  history?: ChatMessageItem[];
}

export interface AiChatResponse {
  response: string;
}

export interface YoutubeInput {
  url: string;
}

export interface YoutubeSummary {
  title: string;
  summary: string;
  keyPoints: string[];
  bulletNotes: string[];
}

export interface GenerateNotesInput {
  text: string;
  title?: string;
}

export interface GeneratedNotes {
  title: string;
  content: string;
  tldr: string;
  keyConcepts: string;
}

export interface GenerateFlashcardsInput {
  content: string;
  count?: number;
}

export interface GenerateQuizInput {
  content: string;
  count?: number;
}

export interface GenerateStudyPlanInput {
  subjects: string;
  examDate: string;
  hoursPerDay?: number;
}

export interface DayPlan {
  day: string;
  date: string;
  tasks: string[];
  hours: number;
}

export interface GeneratedStudyPlan {
  overview: string;
  dailyPlan: DayPlan[];
  tips: string[];
}

export type SimplifyInputLevel =
  (typeof SimplifyInputLevel)[keyof typeof SimplifyInputLevel];

export const SimplifyInputLevel = {
  beginner: "beginner",
  exam: "exam",
  advanced: "advanced",
} as const;

export interface SimplifyInput {
  concept: string;
  level: SimplifyInputLevel;
}

export interface SimplifyResponse {
  explanation: string;
  level: string;
}

export interface AskAboutNoteInput {
  noteContent: string;
  question: string;
}

export type GetNotesParams = {
  search?: string;
};

export type GetFlashcardsParams = {
  noteId?: number;
};