# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StudyFlow AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notes:
    get:
      operationId: getNotes
      tags: [notes]
      summary: List all notes
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Notes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteInput"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── FLASHCARDS ───────────────────────────────────────────────────────────────
  /flashcards:
    get:
      operationId: getFlashcards
      tags: [flashcards]
      summary: List all flashcards
      parameters:
        - in: query
          name: noteId
          schema:
            type: integer
      responses:
        "200":
          description: Flashcards list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flashcard"
    post:
      operationId: createFlashcard
      tags: [flashcards]
      summary: Create a flashcard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFlashcardInput"
      responses:
        "201":
          description: Created flashcard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flashcard"
  /flashcards/{id}:
    patch:
      operationId: updateFlashcard
      tags: [flashcards]
      summary: Update a flashcard (e.g. mark reviewed)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFlashcardInput"
      responses:
        "200":
          description: Updated flashcard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flashcard"
    delete:
      operationId: deleteFlashcard
      tags: [flashcards]
      summary: Delete a flashcard
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── QUIZZES ─────────────────────────────────────────────────────────────────
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Send a message to AI tutor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
  /ai/summarize-youtube:
    post:
      operationId: summarizeYoutube
      tags: [ai]
      summary: Summarize a YouTube video
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YoutubeInput"
      responses:
        "200":
          description: Summary result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YoutubeSummary"
  /ai/generate-notes:
    post:
      operationId: generateNotes
      tags: [ai]
      summary: Generate structured notes from text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateNotesInput"
      responses:
        "200":
          description: Generated notes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedNotes"
  /ai/generate-flashcards:
    post:
      operationId: generateFlashcards
      tags: [ai]
      summary: Generate flashcards from note content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateFlashcardsInput"
      responses:
        "200":
          description: Generated flashcards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlashcardItem"
  /ai/generate-quiz:
    post:
      operationId: generateQuiz
      tags: [ai]
      summary: Generate a quiz from note content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateQuizInput"
      responses:
        "200":
          description: Generated quiz
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizQuestion"
  /ai/generate-study-plan:
    post:
      operationId: generateStudyPlan
      tags: [ai]
      summary: Generate a study plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateStudyPlanInput"
      responses:
        "200":
          description: Generated study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedStudyPlan"
  /ai/simplify:
    post:
      operationId: simplifyConceptAi
      tags: [ai]
      summary: Simplify a concept at different levels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimplifyInput"
      responses:
        "200":
          description: Simplified explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimplifyResponse"
  /ai/ask-about-note:
    post:
      operationId: askAboutNote
      tags: [ai]
      summary: Ask AI a question about a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskAboutNoteInput"
      responses:
        "200":
          description: AI answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ─── NOTE ────────────────────────────────────────────────────────────────────
    Note:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        tags:
          type: string
        tldr:
          type: ["string", "null"]
        keyConcepts:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, content, tags, createdAt, updatedAt]
    CreateNoteInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tags:
          type: string
        tldr:
          type: ["string", "null"]
        keyConcepts:
          type: ["string", "null"]
      required: [title, content]
    UpdateNoteInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tags:
          type: string
        tldr:
          type: ["string", "null"]
        keyConcepts:
          type: ["string", "null"]

    # ─── FLASHCARD ───────────────────────────────────────────────────────────────
    Flashcard:
      type: object
      properties:
        id:
          type: integer
        noteId:
          type: ["integer", "null"]
        question:
          type: string
        answer:
          type: string
        lastReviewed:
          type: ["string", "null"]
        nextReview:
          type: ["string", "null"]
        difficulty:
          type: integer
        createdAt:
          type: string
      required: [id, question, answer, difficulty, createdAt]
    CreateFlashcardInput:
      type: object
      properties:
        noteId:
          type: ["integer", "null"]
        question:
          type: string
        answer:
          type: string
      required: [question, answer]
    UpdateFlashcardInput:
      type: object
      properties:
        lastReviewed:
          type: string
        nextReview:
          type: string
        difficulty:
          type: integer
    FlashcardItem:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
      required: [question, answer]

    # ─── QUIZ ─────────────────────────────────────────────────────────────────────
...
    QuizQuestion:
      type: object
      properties:
        question:
          type: string
        type:
          type: string
          enum: [multiple_choice, true_false]
        options:
          type: array
          items:
            type: string
        answer:
          type: string
        explanation:
          type: string
      required: [question, type, options, answer, explanation]

    # ─── CHAT ─────────────────────────────────────────────────────────────────────
...
    AiChatInput:
      type: object
      properties:
        message:
          type: string
        subject:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessageItem"
      required: [message]
    ChatMessageItem:
      type: object
      properties:
        role:
          type: string
        content:
          type: string
      required: [role, content]
    AiChatResponse:
      type: object
      properties:
        response:
          type: string
      required: [response]
    YoutubeInput:
      type: object
      properties:
        url:
          type: string
      required: [url]
    YoutubeSummary:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        bulletNotes:
          type: array
          items:
            type: string
      required: [title, summary, keyPoints, bulletNotes]
    GenerateNotesInput:
      type: object
      properties:
        text:
          type: string
        title:
          type: string
      required: [text]
    GeneratedNotes:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tldr:
          type: string
        keyConcepts:
          type: string
      required: [title, content, tldr, keyConcepts]
    GenerateFlashcardsInput:
      type: object
      properties:
        content:
          type: string
        count:
          type: integer
      required: [content]
    GenerateQuizInput:
      type: object
      properties:
        content:
          type: string
        count:
          type: integer
      required: [content]
    GenerateStudyPlanInput:
      type: object
      properties:
        subjects:
          type: string
        examDate:
          type: string
        hoursPerDay:
          type: integer
      required: [subjects, examDate]
    GeneratedStudyPlan:
      type: object
      properties:
        overview:
          type: string
        dailyPlan:
          type: array
          items:
            $ref: "#/components/schemas/DayPlan"
        tips:
          type: array
          items:
            type: string
      required: [overview, dailyPlan, tips]
    DayPlan:
      type: object
      properties:
        day:
          type: string
        date:
          type: string
        tasks:
          type: array
          items:
            type: string
        hours:
          type: integer
      required: [day, date, tasks, hours]
    SimplifyInput:
      type: object
      properties:
        concept:
          type: string
        level:
          type: string
          enum: [beginner, exam, advanced]
      required: [concept, level]
    SimplifyResponse:
      type: object
      properties:
        explanation:
          type: string
        level:
          type: string
      required: [explanation, level]
    AskAboutNoteInput:
      type: object
      properties:
        noteContent:
          type: string
        question:
          type: string
      required: [noteContent, question]