# lib/api-zod/src/generated/types/simplifyInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SimplifyInputLevel } from "./simplifyInputLevel";

export interface SimplifyInput {
  concept: string;
  level: SimplifyInputLevel;
}